Added all logic to VM
